+++ packages/protocol/src/exports.ts
@@ -1,1 +1,1 @@
-export * from './utils/exports'
+export { }
