--- packages/protocol/test/unit/exports.test.ts
@@ -1,16 +1,0 @@
-import * as Protocol from '../../src/exports'
-import * as Utils from '../../src/utils/exports'
-
-describe('exports.ts', () => {
-    it('exports all members of containers at top level also ensuring there are no duplicate names', () => {
-        const containers = [Utils]
-        const numKeys = containers.map((value) => Object.keys(value).length).reduce((a, b) => a + b)
-        expect.assertions(numKeys)
-        containers.forEach((container) => {
-            Object.entries(container).forEach(([containerKey, containerValue]) => {
-                // @ts-expect-error figure out proper typing here
-                expect(Protocol[containerKey]).toBe(containerValue)
-            })
-        })
-    })
-})
