+++ packages/utils/src/partition.ts
