+++ packages/utils/test/StreamPartID.test.ts
@@ -1,6 +1,6 @@
-import { toEthereumAddress } from '@streamr/utils'
-import { toStreamID } from '../../../src/utils/StreamID'
-import { StreamPartIDUtils, toStreamPartID } from '../../../src/utils/StreamPartID'
+import { toEthereumAddress } from '../src/EthereumAddress'
+import { toStreamID } from '../src/StreamID'
+import { StreamPartIDUtils, toStreamPartID } from '../src/StreamPartID'
 
 const INVALID_PARTITIONS = [-1, 100, Math.PI, 'abc' as unknown as number]
